requirements.txt
pillow
fpdf2
reportlab